사용자 비밀번호 변경 로직 구현, 요청자와 자원소유권자 동일 여부 확인 기능 구현

+++ src/main/java/com/example/server/member/controller/MemberController.java
@@ -19,7 +19,6 @@ import javax.servlet.http.HttpServletResponse;
 @RequiredArgsConstructor
 public class MemberController {
     private final MemberService memberService;
-    private final TokenService tokenService;
 
     @PostMapping("/success")
     public ResponseEntity success(){
@@ -37,7 +36,7 @@ public class MemberController {
     }
 
     @PostMapping("/logout")
-    public ResponseEntity logout(HttpServletRequest request, HttpServletResponse response){
+    public ResponseEntity logout(HttpServletRequest request){
         String accessToken = request.getHeader("Access-Token");
         String refreshToken = request.getHeader("Refresh-Token");
 
@@ -62,22 +61,47 @@ public class MemberController {
     }
 
     @GetMapping("/{member-id}")
-    ResponseEntity read(@PathVariable("member-id") Long memberId){
+    ResponseEntity read(@PathVariable("member-id") Long memberId, HttpServletRequest request){
+        Long requestId = Long.valueOf((Integer) request.getAttribute("memberId"));
+        if(!memberService.isRequesterSameOwner(requestId, memberId))
+            return new ResponseEntity("요청자와 자원소유자가 다릅니다.", HttpStatus.FORBIDDEN);
+
         MemberResponseDto response = memberService.read(memberId);
 
         return new ResponseEntity(response, HttpStatus.OK);
     }
 
     @PatchMapping("/{member-id}")
-    ResponseEntity update(@RequestBody MemberUpdateDto dto, @PathVariable("member-id") Long memberId){
+    ResponseEntity update(@RequestBody MemberUpdateDto dto, @PathVariable("member-id") Long memberId,
+                          HttpServletRequest request){
+        Long requestId = Long.valueOf((Integer) request.getAttribute("memberId"));
+        if(!memberService.isRequesterSameOwner(requestId, memberId))
+            return new ResponseEntity("요청자와 자원소유자가 다릅니다.", HttpStatus.FORBIDDEN);
+
         Long response = memberService.update(dto, memberId);
 
         if(response == -1) return new ResponseEntity(response, HttpStatus.ACCEPTED);
         return new ResponseEntity(response, HttpStatus.OK);
     }
 
+    @PatchMapping("/password/{member-id}")
+    ResponseEntity updatePassword(@PathVariable("member-id") Long memberId, @RequestBody MemberPasswordUpdateDto dto,
+                                  HttpServletRequest request){
+        Long requestId = Long.valueOf((Integer) request.getAttribute("memberId"));
+        if(!memberService.isRequesterSameOwner(requestId, memberId))
+            return new ResponseEntity("요청자와 자원소유자가 다릅니다.", HttpStatus.FORBIDDEN);
+
+        Long response = memberService.updatePassword(memberId, dto);
+
+        return new ResponseEntity(response, HttpStatus.OK);
+    }
+
     @DeleteMapping("/{member-id}")
-    ResponseEntity delete(@PathVariable("member-id") Long memberId){
+    ResponseEntity delete(@PathVariable("member-id") Long memberId, HttpServletRequest request){
+        Long requestId = Long.valueOf((Integer) request.getAttribute("memberId"));
+        if(!memberService.isRequesterSameOwner(requestId, memberId))
+            return new ResponseEntity("요청자와 자원소유자가 다릅니다.", HttpStatus.FORBIDDEN);
+
         memberService.delete(memberId);
 
         return new ResponseEntity<>(true, HttpStatus.OK);
